requirements.txt
aiofiles==25.1.0
altair==6.0.0
anyio==4.13.0
asgi-csrf==0.11
asgiref==3.11.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-default-group==1.2.4
datasette==0.65.2
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hupper==1.12.1
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mergedeep==1.3.4
narwhals==2.18.1
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
protobuf==6.33.6
pyarrow==23.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sqlparse==0.5.5
streamlit==1.55.0
tabulate==0.10.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
yoyo-migrations==9.0.0
zipp==3.23.0
